origin update notif, cuti detail approval, dll

+++ backend/src/main/java/sys/hris/sims/leave/dto/LeaveBalanceResponse.java
@@ -1,5 +1,6 @@
 package sys.hris.sims.leave.dto;
 
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
@@ -8,7 +9,7 @@ import lombok.Data;
 public class LeaveBalanceResponse {
     private Long employeeId;
     private String employeeName;
-    private Integer annualQuota;
-    private Integer usedAnnualLeave;
-    private Integer remainingAnnualLeave;
+    private BigDecimal annualQuota;
+    private BigDecimal usedAnnualLeave;
+    private BigDecimal remainingAnnualLeave;
 }
